# docs/swagger.yaml
swagger: '2.0'
info:
  title: Living Progress - Build - Mentor Me
  description: Living Progress - Build - Mentor Me API.
  version: "1.0.0"
# the sample made up domain of the service
host: api.mentorme.com
...
schemes:
  - http
  - https
# will be prefixed to all paths
securityDefinitions:
  basicAuth:
    type: basic
# HTTP Basic Authentication. Works over `HTTP` and `HTTPS`
basePath: /v1
produces:
  - application/json
paths:
...
  /mentors/{id}:
    get:
      tags:
        - Mentor
      summary: get Mentor
      description: This endpoint is responsible for getting Mentor
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: number
          required: true
      responses:
        200:
          description: The response containing the Mentor
          schema:
            $ref: '#/definitions/Mentor'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'

    put:
      tags:
        - Mentor
      summary: update Mentor
      description: This endpoint is responsible for updating Mentor
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
        - name: asset
          in: body
          description: The Mentor to update
          schema:
            $ref: '#/definitions/Mentor'
      responses:
        200:
          description: The response containing the Mentor
          schema:
            $ref: '#/definitions/Mentor'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'

    delete:
      tags:
        - Mentor
      summary: delte Mentor
      description: This endpoint is responsible for deleting Mentor
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
      responses:
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /mentors/{id}/averageScore:
      get:
        tags:
          - Mentor
        summary: get Mentor average score
        description: This endpoint is responsible for getting Mentor
        security:
          - basicAuth: []
        parameters:
          - name: id
            in: path
            description: The id
            type: number
            required: true
        responses:
          200:
            description: The response containing the score
            schema:
              type: number
          400:
            description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
            schema:
              $ref: '#/definitions/Error'
          401:
            description: Authentication Error - ien is missing or invalid
            schema:
              $ref: '#/definitions/Error'
          403:
            description: Access Denied Error - if forbidden to access
            schema:
              $ref: '#/definitions/Error'
          404:
            description: Not Found - if match entity not found
            schema:
              $ref: '#/definitions/Error'
          500:
            description: Internal server error - if the request was properly formatted, but the operation failed on the server side
            schema:
              $ref: '#/definitions/Error'
          default:
            description: Unexpected error.
            schema:
              $ref: '#/definitions/Error'
  /mentors/{id}/matchingMentees:
      get:
        tags:
          - Mentor
        summary: get Mentor match mentees
        description: This endpoint is responsible for getting match mentees Mentees
        security:
          - basicAuth: []
        parameters:
          - name: id
            in: path
            description: The id
            type: number
            required: true
        responses:
          200:
            description: The response containing the  match mentees
            schema:
              type: array
              items:
                 $ref: '#/definitions/Mentee'
          400:
            description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
            schema:
              $ref: '#/definitions/Error'
          401:
            description: Authentication Error - ien is missing or invalid
            schema:
              $ref: '#/definitions/Error'
          403:
            description: Access Denied Error - if forbidden to access
            schema:
              $ref: '#/definitions/Error'
          404:
            description: Not Found - if match entity not found
            schema:
              $ref: '#/definitions/Error'
          500:
            description: Internal server error - if the request was properly formatted, but the operation failed on the server side
            schema:
              $ref: '#/definitions/Error'
          default:
            description: Unexpected error.
            schema:
              $ref: '#/definitions/Error'
  /mentors/linkedInExperience:
    get:
      tags:
        - Mentor
      summary: get Mentor linkedInExperience
      description: This endpoint is responsible for getting linkedInExperience for Mentor
      security:
        - basicAuth: []
      responses:
        200:
          description: The response containing the experience data
          schema:
            type: array
            items:
              $ref: '#/definitions/ProfessionalExperienceData'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /mentors:
    get:
      tags:
        - Mentor
      summary: search for entities
      description: This endpoint is responsible for searching by criteria
      security:
      - basicAuth: []
      parameters:
        - name: institutionId
          in: query
          description: The  institution id
          type: number
          required: false
        - name: status
          in: query
          description: The  status
          type: string
          required: false
        - name: minAveragePerformanceScore
          in: query
          description: The  min average performance scores
          type: string
          required: false
        - name: maxAveragePerformanceScore
          in: query
          description: The  max average performance scores
          type: string
          required: false
        - name: name
          in: query
          description: The name
          type: string
          required: false
        - name: personalInterests
          in: query
          type: number
          required: false
          description: should use personalInterests[0]=id
        - name: professionalInterests
          in: query
          type: number
          required: false
          description: should use professional[0]=id
        - name: assignedToInstitution
          in: query
          description: The assigned to institution flag
          type: boolean
          required: false
        - name: mentorType
          in: query
          description: The  mentor type
          type: string
          required: false
        - name: professionalAreas
          in: query
          description: use format professionalAreas[0].id=id
          type: number
          required: false
        - name: companyName
          in: query
          description: The company name
          type: number
          required: false
        - name: mentorRequestStatus
          in: query
          description: The mentor request status
          type: number
          required: false
        - name: pageNumber
          in: query
          description: The page number
          type: number
          format: int32
          required: false
        - name: pageSize
          in: query
          description: The page size
          type: number
          format: int32
          required: false
        - name: sortColumn
          in: query
          description: The sort column
          type: string
          required: false
        - name: sortOrder
          in: query
          description: The sort order
          type: string
          required: false
      responses:
        200:
          description: The response containing the Mentor
          schema:
            $ref: '#/definitions/MentorSearchResult'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
    post:
      tags:
        - Mentor
      summary: create Mentor
      description: This endpoint is responsible for creating Mentor
      security:
        - basicAuth: []
      parameters:
        - name: institution
          in: body
          description: The Mentor to create
          schema:
            $ref: '#/definitions/MentorCreateBody'
      responses:
        200:
          description: The response containing the Institution
          schema:
            $ref: '#/definitions/Mentor'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  UserRole:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      value:
        type: string
        description: The value.
...
  Country:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      value:
        type: string
        description: The value.
  State:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      value:
        type: string
        description: The value.
...
  ParentConsent:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The id.
        parentName:
          type: integer
          format: int32
          description: The parent name.
        signatureFilePath:
          type: string
          description: The signature file path.
        parentEmail:
          type: string
          description: The parent email.
        token:
          type: string
          description: The token.
  InstitutionAffiliationCode:
        type: object
        properties:
          id:
            type: integer
            format: int32
            description: The id.
          code:
            type: string
            description: The code.
          institutionId:
            type: number
            description: The institution id.
          used:
            type: boolean
            description: The used flag.
  PersonalInterest:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The id.
        value:
          type: string
          description: The value.
        picturePath:
          type: string
          description: The picture path.
        parentCategory:
          $ref: '#/definitions/PersonalInterest'
  WeightedPersonalInterest:
        type: object
        properties:
          id:
            type: integer
            format: int32
            description: The id.
          weight:
            type: number
            description: The weight.
          userId:
            type: number
            description: The user id.
          personalInterest:
            $ref: '#/definitions/PersonalInterest'
  ProfessionalInterest:
        type: object
        properties:
          id:
            type: integer
            format: int32
            description: The id.
          value:
            type: string
            description: The value.
          picturePath:
            type: string
            description: The picture path.
          parentCategory:
            $ref: '#/definitions/ProfessionalInterest'
  WeightedProfessionalInterest:
        type: object
        properties:
          id:
            type: integer
            format: int32
            description: The id.
          weight:
            type: number
            description: The weight.
          userId:
            type: number
            description: The user id.
          professionalInterest:
            $ref: '#/definitions/ProfessionalInterest'
...
  Institution:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      createdOn:
        type: integer
        format: int32
        description: The created on datetime with unix epoch format.
      institutionName:
        type: string
        description: The name.
      parentOrganization:
        type: string
        description: The parent organization.
      streetAddress:
        type: string
        description: The street address.
      city:
        type: string
        description: The city.
      zip:
        type: string
        description: The zip.
      phone:
         type: string
         description: The phone.
      email:
        type: string
        description: The email.
      description:
        type: string
        description: The description.
      status:
        type: string
        description: The status.
      logoPath:
        type: string
        description: The logo path.
      contacts:
        type: array
        items:
          $ref: '#/definitions/InstitutionContact'
      state:
        $ref: '#/definitions/State'
      country:
        $ref: '#/definitions/Country'
...
  MentorCreateBody:
    type: object
    properties:
      username:
        type: string
        description: The username.
      firstName:
        type: string
        description: The first name.
      lastName:
        type: string
        description: The last name.
      email:
        type: string
        description: The email.
      profilePicturePath:
        type: string
        description: The profile picture path.
      status:
        type: string
        description: The status.
      assignedToInstitution:
        type: boolean
        description: The assigned to institution flag.
      birthDate:
        type: string
        description: The birth date.
      phone:
         type: string
         description: The phone.
      skypeUsername:
        type: string
        description: The skype username.
      introVideoLink:
        type: string
        description: The intro video link.
      description:
        type: string
        description: The description.
      averagePerformanceScore:
        type: number
        description: The average performance score.
      mentorType:
        type: string
        description: The mentor type.
      companyName:
        type: string
        description: The company name.
      linkedInUrl:
        type: string
        description: The linkedIn url.
      roles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
      personalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedPersonalInterest'
      professionalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedProfessionalInterest'
      institution:
        $ref: '#/definitions/Institution'
      institutionAffiliationCode:
        $ref: '#/definitions/InstitutionAffiliationCode'
      parentConsent:
        $ref: '#/definitions/ParentConsent'
  Mentor:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      createdOn:
        type: integer
        format: int32
        description: The created on datetime with unix epoch format.
      username:
        type: string
        description: The username.
      firstName:
        type: string
        description: The first name.
      lastName:
        type: string
        description: The last name.
      email:
        type: string
        description: The email.
      profilePicturePath:
        type: string
        description: The profile picture path.
      status:
        type: string
        description: The status.
      assignedToInstitution:
        type: boolean
        description: The assigned to institution flag.
      birthDate:
        type: string
        description: The birth date.
      phone:
         type: string
         description: The phone.
      skypeUsername:
        type: string
        description: The skype username.
      introVideoLink:
        type: string
        description: The intro video link.
      description:
        type: string
        description: The description.
      averagePerformanceScore:
        type: number
        description: The average performance score.
      mentorType:
        type: string
        description: The mentor type.
      companyName:
        type: string
        description: The company name.
      linkedInUrl:
        type: string
        description: The linkedIn url.
      roles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
      personalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedPersonalInterest'
      professionalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedProfessionalInterest'
      institution:
        $ref: '#/definitions/Institution'
      institutionAffiliationCode:
        $ref: '#/definitions/InstitutionAffiliationCode'
      parentConsent:
        $ref: '#/definitions/ParentConsent'
  MentorSearchResult:
    type: object
    properties:
      total:
        type: number
        format: integer
        description: The total.
      totalPages:
        type: number
        format: integer
        description: The total pages.
      entities:
        type: array
        items:
          $ref: '#/definitions/Mentor'
...
  Mentee:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      createdOn:
        type: integer
        format: int32
        description: The created on datetime with unix epoch format.
      username:
        type: string
        description: The username.
      firstName:
        type: string
        description: The first name.
      lastName:
        type: string
        description: The last name.
      email:
        type: string
        description: The email.
      profilePicturePath:
        type: string
        description: The profile picture path.
      status:
        type: string
        description: The status.
      assignedToInstitution:
        type: boolean
        description: The assigned to institution flag.
      birthDate:
        type: string
        description: The birth date.
      phone:
         type: string
         description: The phone.
      skypeUsername:
        type: string
        description: The skype username.
      introVideoLink:
        type: string
        description: The intro video link.
      description:
        type: string
        description: The description.
      averagePerformanceScore:
        type: number
        description: The average performance score.
      familyIncome:
        type: number
        description: The family income.
      school:
        type: string
        description: The school.
      facebookUrl:
        type: string
        description: The facebook url.
      roles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
      personalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedPersonalInterest'
      professionalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedProfessionalInterest'
      institution:
        $ref: '#/definitions/Institution'
      institutionAffiliationCode:
        $ref: '#/definitions/InstitutionAffiliationCode'
      parentConsent:
        $ref: '#/definitions/ParentConsent'
...
  ProfessionalExperienceData:
    type: object
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string